extracted implementation part of `Expr.h` to `Expr.c`
